Update metadata for eia_bulk_elec to eiaapi (#4212)
* Update metadata for eia_bulk_elec to eia_bulk

* Fix index error and update source name

* Rename eia_bulk_elec to eia_bulk

* Rename eia_bulk to eiaapi

* Update release notes
