+++ dbt/models/eiaapi/core_eia__yearly_fuel_receipts_costs_aggs/schema.yml
