+++ src/pudl/extract/__init__.py
@@ -18,8 +18,8 @@ from . import (
     eia861,
     eia923,
     eia930,
-    eia_bulk_elec,
     eiaaeo,
+    eiaapi,
     epacems,
     excel,
     ferc1,
