+++ src/pudl/transform/__init__.py
@@ -70,8 +70,8 @@ from . import (
     eia861,
     eia923,
     eia930,
-    eia_bulk_elec,
     eiaaeo,
+    eiaapi,
     epacems,
     ferc,
     ferc1,
